fixat ownerStuff

+++ index.php
@@ -10,10 +10,9 @@ require_once("cars_model.php");
 require_once("response.php");
 require_once("auth.php");
 
-app::get("/session", function(){
+app::get("/session", function () {
 
     var_dump($_SESSION);
-
 });
 
 
@@ -23,17 +22,23 @@ app::get('/error/$message', "html/error");
 
 app::get("/cars", "html/cars");
 app::get("/cars/create", "html/createCar.php");
-app::post("/cars/create", function(){
+app::post("/cars/create", function () {
+
+    if (!Auth::checkAuth()) {
+        Res::redirect("/login?error=not authenticated");
+        return;
+    }
+
     $check = true;
     $message = "";
-    foreach($_POST as $key => $input){
-        if(empty(trim($input))) {
+    foreach ($_POST as $key => $input) {
+        if (empty(trim($input))) {
             $check = false;
             $message = $key . " is required";
             break;
         }
     }
-    if(!$check) {
+    if (!$check) {
         res::redirect("/error/$message");
         return;
     }
@@ -42,19 +47,36 @@ app::post("/cars/create", function(){
 });
 
 
-app::post("/cars/update", function(){
+app::post("/cars/update", function () {
 
-    Cars::updateCar($_POST);
-    Res::redirect("/cars");
+    if (!Auth::checkAuth()) {
+        Res::redirect("/login?error=not authenticated");
+        return;
+    }
 
+    try {
+        Cars::updateCar($_POST);
+        Res::redirect("/cars");
+    } catch (Exception $e) {
+        Res::redirect("/cars?error=".$e->getMessage());
+    }
 });
 
-app::get('/deletecar/$id', function($id){
-    Cars::deleteCar($id);
-    Res::redirect("/cars");
+app::get('/deletecar/$id', function ($id) {
+
+    if (!Auth::checkAuth()) {
+        Res::redirect("/login?error=not authenticated");
+        return;
+    }
+    try {
+        Cars::deleteCar($id);
+        Res::redirect("/cars");
+    } catch (Exception $e) {
+        Res::redirect("/cars?error=" . $e->getMessage());
+    }
 });
 
-app::get('/cars/search/$name', function($name){
+app::get('/cars/search/$name', function ($name) {
     echo "CAR width name: $name";
 });
 
@@ -63,44 +85,41 @@ app::get('/cars/search/$name', function($name){
 // auth-routes
 
 app::get("/register", "html/register");
-app::post("/register", function(){
+app::post("/register", function () {
 
-$email = $_POST['email'] ?? "";
-if(!filter_var($email, FILTER_VALIDATE_EMAIL)){
-    Res::redirect("/register?error=email error");
-    return;
-}
-$password = $_POST['password'] ?? "";
-if(strlen($password)<8){
-    Res::redirect("/register?error=Password must be at least 8 char long");
-    return;
-}
-$user = [
-    "role"=>"user",
-    "email"=>$email,
-    "password"=>$password
-];
+    $email = $_POST['email'] ?? "";
+    if (!filter_var($email, FILTER_VALIDATE_EMAIL)) {
+        Res::redirect("/register?error=email error");
+        return;
+    }
+    $password = $_POST['password'] ?? "";
+    if (strlen($password) < 8) {
+        Res::redirect("/register?error=Password must be at least 8 char long");
+        return;
+    }
+    $user = [
+        "role" => "user",
+        "email" => $email,
+        "password" => $password
+    ];
 
-$regReturn = Auth::register($user);
-if(isset($regReturn['error'])){
-  return  Res::redirect("/login?error=".$regReturn['error']);
-}
+    $regReturn = Auth::register($user);
+    if (isset($regReturn['error'])) {
+        return  Res::redirect("/login?error=" . $regReturn['error']);
+    }
 
-Res::redirect("/login?message=Please Login");
-
+    Res::redirect("/login?message=Please Login");
 });
 
 
-app::post("/login", function(){
+app::post("/login", function () {
 
-   try{
-    $loginReturn = Auth::login($_POST);
-    Res::redirect("/cars?message=$loginReturn");
-   } 
-   catch(Exception $e){
-    Res::redirect("/login?error=".$e->getMessage());
-   }
-
+    try {
+        $loginReturn = Auth::login($_POST);
+        Res::redirect("/cars?message=$loginReturn");
+    } catch (Exception $e) {
+        Res::redirect("/login?error=" . $e->getMessage());
+    }
 });
 
-app::get("/login", "html/login" );
+app::get("/login", "html/login");
